Fix launch process of Mistral components
* Fixed the bug #1622534
* Introduced MistralService class extending oslo.service.Service
  that all Mistral components running standalone should extend
* Refactored engine, executor and event engine with MistralService
* Moved most of the startup logic from launch.cmd to individual
  corresponding components
* Cluster membership is now under control of MistralService
* Fixed test_join.test_full_join_with_conditions() which previously
  had a bug and we were just 'lucky' that it passed due to a
  different work of scheduler
* Fixed a number of test cases
* Other minor changes

TODO:
* We now use many launchers (from oslo.service) to launch
  services whereas we could use just one which is recommended by
  oslo. However, we can't do that because of the api service
  that uses many workers. We'll need to look at how to refactor it
  moving forward.
* Write tests for MistralService and its derrived classes
* Address a number of TODO comments.

Closes-Bug: 1622534

+++ mistral/db/v2/sqlalchemy/api.py
@@ -16,6 +16,7 @@
 
 import contextlib
 import sys
+import threading
 
 from oslo_config import cfg
 from oslo_db import exception as db_exc
@@ -42,6 +43,10 @@ CONF = cfg.CONF
 LOG = logging.getLogger(__name__)
 
 
+_SCHEMA_LOCK = threading.RLock()
+_initialized = False
+
+
 def get_backend():
     """Consumed by openstack common code.
 
@@ -52,20 +57,33 @@ def get_backend():
 
 
 def setup_db():
-    try:
-        models.Workbook.metadata.create_all(b.get_engine())
-    except sa.exc.OperationalError as e:
-        raise exc.DBError("Failed to setup database: %s" % e)
+    global _initialized
+
+    with _SCHEMA_LOCK:
+        if _initialized:
+            return
+
+        try:
+            models.Workbook.metadata.create_all(b.get_engine())
+
+            _initialized = True
+        except sa.exc.OperationalError as e:
+            raise exc.DBError("Failed to setup database: %s" % e)
 
 
 def drop_db():
-    global _facade
+    global _initialized
 
-    try:
-        models.Workbook.metadata.drop_all(b.get_engine())
-        _facade = None
-    except Exception as e:
-        raise exc.DBError("Failed to drop database: %s" % e)
+    with _SCHEMA_LOCK:
+        if not _initialized:
+            return
+
+        try:
+            models.Workbook.metadata.drop_all(b.get_engine())
+
+            _initialized = False
+        except Exception as e:
+            raise exc.DBError("Failed to drop database: %s" % e)
 
 
 # Transaction management.
